requirements.txt
numpy
tensorflow
keras
opencv-python
tqdm
matplotlib
python-dotenv
